zip file extractor and pdf merger created
Version 1:
The Zip File Extractor uses ZipFile to extract all the zip files in a folder. The program is currently made to extract all the zips files in a sub folder. 
The PDF Merger then reads each folder and its contents, and merges the files together and adds them to a different directory. 

Future updates:
Will be adding functions to handle the application differently. (e.g unzip and merge one object)